requirements.txt
numpy
scipy
PyYAML
pandas
wandb
scikit-learn
networkx
tqdm
fairseq
matplotlib
seaborn